+++ test/Dapr.Jobs.Test/Extensions/StringExtensionsTests.cs
@@ -2,7 +2,7 @@
 using Dapr.Jobs.Extensions;
 using Xunit;
 
-namespace Dapr.Jobs.Test;
+namespace Dapr.Jobs.Test.Extensions;
 
 public class StringExtensionsTests
 {
